requirements.txt
# Core dependencies
python-dotenv>=0.19.0
openai==0.28.0
alpaca-py>=0.8.0

# Data handling
pandas>=1.3.0
numpy>=1.21.0
pandas-ta==0.3.14b
yfinance==0.2.31

# API Connections
websocket-client==1.6.0
requests>=2.26.0
aiohttp==3.9.1

# Discord Integration
discord-webhook==1.3.0

# Utilities
pytz==2023.3
schedule==1.2.0
python-dateutil>=2.8.2

# Logging
colorlog==6.7.0
